// dashboard/src/features/orgs/projects/database/dataGrid/components/ColumnAutocomplete/types.ts
export interface AutocompleteOption<TValue = string> {
  /**
   * Label to display.
   */
  label: string;
  /**
   * Label to display in the dropdown.
   */
  dropdownLabel?: string;
  /**
   * Value to be submitted.
   */
  value: TValue;
  /**
   * Determines whether the option is custom.
   */
  custom?: boolean;
  /**
   * Value that can be used to group options.
   */
  group?: string;
  /**
   * Any additional data to be passed to the option.
   */
  // biome-ignore lint/suspicious/noExplicitAny: option metadata is intentionally untyped
  metadata?: any;
}